Update Hud,add ClickGui.

+++ src/net/minecraft/client/gui/FontRenderer.java
@@ -379,7 +379,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, int x, int y, int color)
+    public int drawString(String text, float x, float y, int color)
     {
         return this.drawString(text, (float)x, (float)y, color, false);
     }
